move Logos to the component
